feat: implement Header and Button components with variants and update exports
