+++ src/components/ui/index.ts
@@ -1,6 +1,8 @@
 export * from './avatar'
 export * from './badge'
 export * from './button'
-export * from './icon-button'
+export * from './heading'
 export * from './inline'
 export * from './input'
+export * from './stack'
+export * from './text'
