+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from '@/components/ui/avatar'
 export * from '@/components/ui/button'
 export * from '@/components/ui/heading'
-export * from '@/components/ui/icon-button'
 export * from '@/components/ui/inline'
 export * from '@/components/ui/stack'
 export * from '@/components/ui/text'
